move more general code from analysis into digicampipe
