+++ digicampipe/scripts/pulse_shape.py
@@ -17,137 +17,14 @@ from digicampipe.io.event_stream import calibration_event_stream
 from tqdm import tqdm
 from docopt import docopt
 
-import numba
-
-
-@numba.jit
-def estimate_arrival_time(adc, thr=0.5):
-    '''
-    estimate the pulse arrival time, defined as the time the leading edge
-    crossed 50% of the maximal height,
-    estimated using a simple linear interpolation.
-
-    *Note*
-    This method breaks down for very small pulses where noise affects the
-    leading edge significantly.
-    Typical pixels have a ~2LSB electronics noise level.
-    Assuming a leading edge length of 4 samples
-    then for a typical pixel, pulses of 40LSB (roughly 7 p.e.)
-    should be fine.
-
-    adc: (1296, 50) dtype=uint16 or so
-    thr: threshold, 50% by default ... can be played with.
-
-    return:
-        arrival_time (1296) in units of time_slices
-    '''
-    n_pixel = adc.shape[0]
-    arrival_times = np.zeros(n_pixel, dtype='f4')
-
-    for pixel_id in range(n_pixel):
-        y = adc[pixel_id]
-        y -= y.min()
-        am = y.argmax()
-        y_ = y[:am+1]
-        lim = y_[-1] * thr
-        foo = np.where(y_ < lim)[0]
-        if len(foo):
-            start = foo[-1]
-            stop = start + 1
-            arrival_times[pixel_id] = start + (
-                (lim-y_[start]) / (y_[stop]-y_[start])
-            )
-        else:
-            arrival_times[pixel_id] = np.nan
-    return arrival_times
-
-
-class Histogram2d:
-
-    def __init__(self, shape, range):
-        self.histo = np.zeros(shape, dtype='u2')
-        self.range = range
-        self.extent = self.range[0] + self.range[1]
-
-    def fill(self, x, y):
-        for pixel_id in range(len(x)):
-            H, xedges, yedges = np.histogram2d(
-                x[pixel_id],
-                y[pixel_id],
-                bins=self.histo.shape[1:],
-                range=self.range
-            )
-            self.histo[pixel_id] += H.astype('u2')
-
-    def contents(self):
-        return self.histo
-
-
-class Histogram2dChunked:
-
-    def __init__(self, shape, range, buffer_size=1000):
-        self.histo = np.zeros(shape, dtype='u2')
-        self.range = range
-        self.extent = self.range[0] + self.range[1]
-
-        self.buffer_size = buffer_size
-        self.buffer_x = None
-        self.buffer_y = None
-        self.buffer_counter = 0
-
-    def __reset_buffer(self, x, y):
-        self.buffer_x = np.zeros(
-            (self.buffer_size, *x.shape),
-            dtype=x.dtype
-        )
-        self.buffer_y = np.zeros(
-            (self.buffer_size, *y.shape),
-            dtype=y.dtype
-        )
-        self.buffer_counter = 0
-
-    def fill(self, x, y):
-        if self.buffer_counter == self.buffer_size:
-            self.__fill_histo_from_buffer()
-
-        if self.buffer_x is None:
-            self.__reset_buffer(x, y)
-
-        self.buffer_x[self.buffer_counter] = x
-        self.buffer_y[self.buffer_counter] = y
-        self.buffer_counter += 1
-
-    def __fill_histo_from_buffer(self):
-        if self.buffer_x is None:
-            return
-
-        self.buffer_x = self.buffer_x[:self.buffer_counter+1]
-        self.buffer_y = self.buffer_y[:self.buffer_counter+1]
-        for pixel_id in range(self.buffer_x.shape[1]):
-            foo = self.buffer_x[:, pixel_id].flatten()
-            bar = self.buffer_y[:, pixel_id].flatten()
-            H, xedges, yedges = np.histogram2d(
-                foo,
-                bar,
-                bins=self.histo.shape[1:],
-                range=self.range
-            )
-            self.histo[pixel_id] += H.astype('u2')
-        self.buffer_x = None
-        self.buffer_y = None
-        self.buffer_counter = 0
-
-    def contents(self):
-        self.__fill_histo_from_buffer()
-        return self.histo
+from digicampipe.calib.camera.time import estimate_arrival_time
+from digicampipe.utils.hist2d import Histogram2dChunked
 
 
 def main(outfile_path, input_files=[]):
     events = calibration_event_stream(input_files)
     Rough_factor_between_single_pe_amplitude_and_integral = 21 / 5.8
     histo = None
-
-    _range = [[-10, 40], [-0.2, 1.5]]
 
     for e in tqdm(events):
         adc = e.data.adc_samples
@@ -173,7 +50,7 @@ def main(outfile_path, input_files=[]):
         if histo is None:
             histo = Histogram2dChunked(
                 shape=(adc.shape[0], 101, 101),
-                range=_range
+                range=[[-10, 40], [-0.2, 1.5]]
             )
 
         histo.fill(
